fixed document editor to follow conventions
