+++ packages/editor/document-editor/src/hooks/use-editor-markings.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { IMarking } from "..";
+import { IMarking } from "src/types/editor-types";
 
 export const useEditorMarkings = () => {
   const [markings, setMarkings] = useState<IMarking[]>([]);
