+++ packages/editor/document-editor/src/types/editor-types.ts
@@ -5,3 +5,9 @@ export interface DocumentDetails {
   last_updated_by: string;
   last_updated_at: Date;
 }
+export interface IMarking {
+  type: "heading";
+  level: number;
+  text: string;
+  sequence: number;
+}
