+++ packages/editor/document-editor/src/ui/extensions/widgets/issue-embed-widget/index.tsx
@@ -1,5 +1,5 @@
-import { IssueWidget } from "./issue-widget-node";
-import { IIssueEmbedConfig } from "./types";
+import { IssueWidget } from "src/ui/extensions/widgets/issue-embed-widget/issue-widget-node";
+import { IIssueEmbedConfig } from "src/ui/extensions/widgets/issue-embed-widget/types";
 
 interface IssueWidgetExtensionProps {
   issueEmbedConfig?: IIssueEmbedConfig;
