+++ packages/editor/document-editor/src/utils/menu-actions.ts
